pyproject.toml
[project]
name = "nord-higiene-web"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "load-dotenv>=0.1.0",
    "flask>=3.0.0",
    "flask-socketio>=5.0.0",
    "python-dotenv>=1.0.0",
    "flask-sqlalchemy>=3.0.0",
    "cryptography>=40.0.0",
    "celery>=5.0.0",
    "openpyxl>=3.0.0",
    "redis>=4.0.0",
    "pandas>=2.0.0",
    "click>=8.3.1",
    "requests>=2.32.5",
    "selenium>=4.41.0",
    "webdriver-manager>=4.0.2",
    "beautifulsoup4>=4.14.3",
]
